# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "S360°",
        "description": "Documentation interactive de S360°, couvrant les modules migrés depuis S360.",
        "contact": {
            "name": "Équipe S360°",
            "email": "support@s360.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Instance locale de developpement"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/audit-logs": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Lister les journaux d audit",
                "description": "Recupere une liste paginee des journaux d audit avec possibilite de filtrage par action, utilisateur, type de modele, dates et recherche. Supporte le groupement par date.",
                "operationId": "listAuditLogs",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d elements par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "description": "Filtrer par action (peut etre multiple, separees par virgule)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "create,update,delete"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrer par ULID de l utilisateur",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$"
                        }
                    },
                    {
                        "name": "model_type",
                        "in": "query",
                        "description": "Filtrer par type de modele (ex: User, Role, Permission)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "User"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Date de debut (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Date de fin (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans la description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_by_date",
                        "in": "query",
                        "description": "Grouper les resultats par date",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginee des journaux d audit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "150 journal(aux) d audit trouve(s)"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit-logs/{id}": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Afficher un journal d audit",
                "description": "Recupere les details complets d un journal d audit specifique, incluant les informations sur l utilisateur associe.",
                "operationId": "showAuditLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du journal d audit",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details du journal d audit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Journal d audit charge"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Journal d audit non trouve"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Utiliser un token d'accès généré via Sanctum.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}